Worked through the unit testing and added/changed infrastructure to correct errors in message testing, screen data page and page processor testing.  Additional previously existing testing gaps (where the test suite exists) have been closed.
Updated some std::formatters for device types and message types which enabled deduplication of printing some messages; this is a corrected pattern that should be used across the codebase.

Removed the Message Loop Start message as it appears to be a collision with the PDA Highlight message.  A side effect of unit test correction is the implementation of correct serialization and deserialisation across the rest of the (tested) messages...note that most messages remain untested.

Fixed some defects in the scraping for the OneTouch device so it can now accurate scape when processing pages.

Refactored the configuration classes so that "devices" are all represented in a tree structure (and are enumerable).