Significant refactoring to separate the "Config" namespace into the "Kernel" namespace and start splitting it away from being Jandy-specific.
Replaced the "Config" with a centralised data hub which can be leveraged by its various dependencies such as web routes without inner knowledge of the Jandy Equipment.

Added a centralised statistics hub as the location for common statistics and other information to be tracked (for reporting purposes).

Replaced boost::regex with re2.