Replace bustache with mstch.
Remove any reference to the bustache repo and replace it with the vcpkg for mstch.  Update all the web routes that render pages to use mstch (including supporting functions).

Fix each of the routes to have external linkage for their route definitions (vs. internal) by marking the route definition as inline.  This fixes a warning under gcc that the route has internal linkage (which could indicate an ODR issue).