Add fix for lifetime of web page routes.
Web page routes were going out of scope early and so the contents of the underlying classes were being destroyed prior to use within the web request handler.  All routes are now derived from a single (additional) interface and are captured in a common collection in the main thread of execution to assure their lifetimes.