Refactor the way devices are created by the OneTouch and the Serial Adapter devices both leverage the JandyAuxillaryFactory.
Integrate the data reported by the Serial Adapter into the DataHub via the message and command processors.

Fix a defect in the cutoffs for temperature minimums (as comparing floats may not always work due to floating point equality).  Fix a defect in the formatting string when localising temperatures in the dashboards.