Refactoring of the data hub events (and websocket events) to add a system service event type.
Add support for parsing the label aux list in the OneTouch device screens to get custom labels for devices.  Add a new start-up screen scraping script and initialise using that.

Add a test modal banner to the index template and start the plumbing for the service event types.

Refactor the static directory and assets locations to remove the redundant styles.css and associated files.

Add support for code coverage under Linux using gcov/lcov (and include a custom debug build as a user preset.