Improve testability of HTTP sessions and WebSocket sessions by consolidating common behaviours into the base ISession interface.
Add a unique session id to each session object so that its changes and actions can be tracked.  Update HTTP session and WebSocket session test cases so that they are a) using the ISession interface and b) are no longer commented out.

Add a mock basic_stream which is derived from beast::test::stream to enable unit testing.

Improve the IWebSocketBase interface so that it closes any active sessions that are orphaned when it is being destroyed.  Additionally, remove the redundant WebSocket_MakeSession templated function.

Abstract the unit test support functions into a TestResponseObject which holds the various mechanisms to connect the unit test to the HTTP server objects under test.